Added new Frim using the new AI code.
Added AI node to wait until the sprite got to an animation frame before continueing.
Added DisableCollision ai node,
fixed dash ai node. Now it properly handles speed and added to it the ability to use the target's position as a relative distance.
Created new frim with everything the old frim had except for the ballistics because it didn't seem to be working properly
fixed foolish bug with knockback where you were always getting pushed away from your target instead of the source of the knockback.